requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.13
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.2
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
